requirements.txt
# === Core ML ===
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
peft>=0.7.0            # LoRA for parameter-efficient fine-tuning (fits 4-6GB VRAM)
bitsandbytes>=0.41.0   # 8-bit quantization

# === Code Analysis ===
# ast is stdlib, no install needed
radon>=6.0.0           # cyclomatic complexity, maintainability index

# === Test Execution & Coverage ===
pytest>=7.0.0
pytest-cov>=4.0.0      # coverage plugin for pytest
coverage>=7.0.0
mutmut>=2.4.0          # mutation testing for reward signal

# === RL (end-term) ===
# trl>=0.7.0           # uncomment when starting RL phase
# pykan>=0.2.0         # uncomment when starting KAN phase

# === Data & Utilities ===
pandas>=2.0.0
tqdm>=4.65.0
GitPython>=3.1.0       # for mining GitHub repos
jsonlines>=4.0.0
scikit-learn>=1.3.0

# === Evaluation ===
nltk>=3.8.0            # BLEU score
codebleu>=0.7.0        # CodeBLEU metric
